requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
kaggle>=1.5.16
PyYAML>=6.0
grad-cam>=1.4.8
umap-learn>=0.5.3
scipy>=1.11.0
